backend/requirements.txt
# Django Core
Django==5.0.0
django-ninja==1.1.0
djangorestframework==3.14.0
email-validator==2.1.0  # Required for pydantic EmailStr validation

# Authentication & OAuth2
django-allauth[socialaccount]==0.59.0
PyJWT==2.8.0

# Async & Celery
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
flower==2.0.1

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# CORS
django-cors-headers==4.3.1

# Environment
python-decouple==3.8

# HTTP & Utilities
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
Pillow==10.2.0
pytesseract==0.3.10

# Production
gunicorn==21.2.0
uvicorn[standard]==0.25.0
whitenoise==6.6.0

# AI & RAG
openai==1.55.3
anthropic==0.75.0
google-generativeai==0.8.3
rank-bm25==0.2.2
sentence-transformers==5.1.2
cohere==5.20.0
voyageai==0.3.6
pgvector==0.4.2

# Charts & Analytics
matplotlib==3.8.2
seaborn==0.13.1

# Report Generation
reportlab==4.0.7
python-docx==1.1.0
